+++ attributes/default.rb
@@ -13,7 +13,7 @@ default["statsd"]["delete_sets"]                  = false
 default["statsd"]["delete_counters"]              = false
 default["statsd"]["username"]                     = "statsd"
 default["statsd"]["config"]                       = {}
-default["statsd"]["plugins"]["enabled"]           = []
+default["statsd"]["backends"]["enabled"]          = []
 
 # Backends default configuration
 default["statsd"]["graphite_host"]                = "127.0.0.1"